+++ workspace/backend/mpick/src/main/java/com/k9c202/mpick/global/entity/TradeImageEntity.java
@@ -5,7 +5,6 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 import java.sql.Timestamp;
 
 import org.hibernate.annotations.CreationTimestamp;
@@ -20,7 +19,7 @@ public class TradeImageEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger tradeImageId;
+    private Long tradeImageId;
 
     @Column
     private Integer tradeId;
